cache renewal: keep sub-second precision of the invalidation time
RenewCacheJob truncated its reference time to whole seconds when
serializing. A cache entry written within the same second as the
change that should invalidate it then compared as "not older" and
wrongly survived — visible as stale aktivitaetszahlen after status
changes, and as the long-standing flakiness of the corresponding
specs, which pass or fail depending on whether a second boundary
happens to fall between computation and change.

https://github.com/fiedl/wingolfsplattform/issues/113

+++ your_platform/app/jobs/renew_cache_job.rb
@@ -5,7 +5,10 @@ class RenewCacheJob < ApplicationJob
   queue_as :cache
 
   def serialize
-    arguments.last[:time] = arguments.last[:time].to_i if arguments.last && arguments.last[:time]
+    # Keep sub-second precision: with whole seconds, cache entries
+    # written in the same second as the change compare as "not older"
+    # and wrongly survive the renewal.
+    arguments.last[:time] = arguments.last[:time].to_f if arguments.last && arguments.last[:time]
     super
   end
